Add accessibility label for center stage. Refactor.
